+++ FirstCommandApp/FirstCommandApp/Rectangle.m
@@ -14,14 +14,14 @@
 }
 @synthesize width, height;
 
--(void) setWidth: (int) w andHeight: (int) h {
+-(void) setWidth: (float) w andHeight: (float) h {
     width = w;
     height = h;
 }
--(int) area {
+-(float) area {
     return width * height;
 }
--(int) perimeter {
+-(float) perimeter {
     return (width + height) * 2;
 }
 -(void) setOrigin: (XYPoint *) pt {
