+++ FirstCommandApp/FirstCommandApp/Square.h
@@ -9,6 +9,6 @@
 #import "Rectangle.h"
 
 @interface Square : Rectangle
--(void) setSide: (int) s;
--(int) side;
+-(void) setSide: (float) s;
+-(float) side;
 @end
