+++ FirstCommandApp/FirstCommandApp/XYPoint.h
@@ -10,8 +10,8 @@
 
 @interface XYPoint : NSObject
 
-@property int x, y;
+@property float x, y;
 
--(void) setX: (int) xVal andY: (int) yVal;
+-(void) setX: (float) xVal andY: (float) yVal;
 
 @end
